requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
dash==3.1.1
fastapi==0.116.1
firebase-admin==6.6.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
git-filter-repo==2.47.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.3
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==21.2.0
h11==0.16.0
h5py==3.14.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.1.0
narwhals==1.47.0
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.0.3
Pillow==10.0.0
plotly==6.2.0
proto-plus==1.25.0
protobuf==4.25.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.5.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.6
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.4.0
rich==14.1.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.35.0
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
